Fix webpack config for projects using "full control mode"
I'm working with a project using the `full control mode` webpack config
option from Storybook.
https://storybook.js.org/configurations/custom-webpack-config/#full-control-mode

We're running into an issue where the resulting webpack config can't be
used for happo. I believe the issue might be because we're sending in
the wrong base webpack config to extend. Instead of sending in the happo
webpack config, we should pass in the storybook base config, then merge
the two after the fact. This brings the logic between extend mode and
full control mode down to the same basic operations:

1. let storybook do its thing
2. pick out the webpack config options that we care about and merge it
with the happo config.