Remove dependency on React
So far, this plugin has only been able to render storybook stories
written in React. This limitation was mostly there because all current
users of the plugin were all using React. We can break out of this
dependency by using the event system provided by Storybook.

This change makes the plugin only work for Storybook v4 and later.

The biggest issue I had was around unmount failures. If a component
throws an error on being removed from the DOM, the failure message
ended up on the next rendered example. To work around that, we now
re-render the story if we see an error message after the first render.
This seems to do the trick.