Prevent large stories from breaking subsequent ones
In 2.1.0, I introduced a regression that caused screenshots to have the
wrong size. The happo workers inject a div for each example equal to the
size of the content. If we don't clean out that div, the next example
will contain measurements from the measuring div. We should fix this on
happo workers as well, but at least this will unbreak the storybook
plugin.