Make a more careful dom cleanup
In 2.1.1, I added a cleanup method to make sure that the size of the
screenshot ended up being right. This led to an issue where if you had
components rendering directly into the dom (such as react portals,
tether.js) the cleanup would be a little too aggressive - leading to
issues when a component was to be unmounted.

Manually cleaning out `#__snapshot-box` isn't ideal since it's really an
implementation detail of the happo.io implementation. But for now it
will get the job done. I've made a note about changing this on the
happo.io side instead.