Wait for data being fetched
Potential fix for #15

To prevent spuriousness coming from the fact that a component can make
asynchronous calls for data, I'm adding a global listener for "xhr
silence", i.e. when no outstanding requests are active.

While we may not catch all data fetching this way, overriding both
xhrrequest and window.fetch should be enough to cover most cases.

For now I'm keeping this code local to this plugin. Going forward, it's
likely that this will be merged into the code that run on happo workers
instead.