Avoid babel-ifying register.js in tests
This file is already processed by babel. Storybook v5 added code that
avoids processing node_modules, and we just need to treat this file the
same way.

In users' codebases, the register.js file will be unprocessed by babel
by storybook since it will be located in node_modules. So this is for
testing purposes only.