Move config.js to preview.js
This is a first step towards using declarative configuration for
Storybook. No code changes made so far, I've simply moved things from
config.js to preview.js.

This commit serves as a POC that happo-plugin-storybook can be used with
Storybook v6. Docs need to be updated to reflect the fact that the
happo-plugin-storybook/register import should go in preview.js. I'll
make a TODO about making that docs change later.