Add `afterScreenshot` hook
Same as for beforeScreenshot, this hook can be used to clean up things
after a screenshot has been taken.

At first I wanted to have the happo worker invoke this call, but I
decided that it was premature. So for now the afterScreenshot hook is
called right before rendering the next story (which technically is
after the screenshot).