Allow before/afterScreenshot hooks to fail
To prevent failing the entire test suite, we can simply log something
and continue execution.

I thought about adding an `await` before the hook calls as well, but
decided it was YAGNI. Plus, we would have to have some kind of timeout
for these calls, and I didn't want to introduce that complexity atm.