Add support for `targets` configuration
Similar to how we do things for Happo Examples and Cypress integrations,
this commit adds support for a `targets` option to stories.

https://docs.happo.io/docs/examples#limiting-targets
https://docs.happo.io/docs/cypress#selecting-targets

The implementation here ended up a little bit hacky. Happo workers are
already communicating the target name to the test suite, but only for
Happo Examples, which uses a `window.happoProcessor` adapter to render
examples. We can (ab)use this protocol by temporarily defining the
`init` method, then quickly cleaning it out so that it won't continue to
be used by Happo workers.

This will need a clenup at some point, probably by adding another `init`
call on the `window.happo` global, possibly in the same realm as
`initChunk`. But for now this should do the trick.

Implements #64