Switch to CSF format, use storyStoreV7
"Story store v7" is a way to lazily load stories to speed up a storybook
project. With this flag enabled, Happo would fail with

Error: Cannot call extract() unless you call cacheAllCSFFiles() first

I've moved some stuff around so that we can call that function. It is
asynchronous, so I couldn't use one of the `init`* functions we already
had defined (they are assumed to be synchronous). The good thing is that
both initChunk and happoProcessor.init are deprecated, `happo.init` is
the way to go.