Move hook invocation to decorator
I'm taking a new approach to calling beforeScreenshot etc in the
preview document. Instead of relying on the register module to set
things up I'm making use of a decorator. The decorator has the local
state needed to invoke functions. I couldn't use `useChannel` here so I
went with a custom useEffect to set things up.