Adapt for Storybook v7
The new version of Storybook has been cooking for about a year and
judging by the activity lately it is about to drop soon. Taking the time
to adapt happo-plugin-storybook for Storybook v7 means we're in a better
position when the upgrade is released.

Most of the work in this PR was already done by @juliepagano in this
Perses PR:
https://github.com/perses/perses/pull/991

I'm trying to be careful enough to keep support for Storybook v6 and v5
and at the same time be compatible with v7. If I can't get Storybook v5
to behave under the new setup, I'll likely drop support for it. We'll
see where this ends up. :)