Stop getting addons channel from @storybook/addons
We're seeing issues where the addons channel exported by
@storybook/addons is different from the globally available channel
window.__STORYBOOK_ADDONS_CHANNEL__.

In the past, we used the global variable but switched to importing it in
05328d7a3de4b668f. I can't recall the details there but seems safe to
move back.

I have manually checked that this fixes things for one of our customer.
The symptoms were that all stories appeared to be rendered but they
didn't actually render on screen. Switching to the global channel did
the trick.