Improve build command detection
If you're using yarn 2, the `yarn list` command won't work. To make
things a little bit more robust, I'm adding a second SB6 detection
command. This time we're simply trying to get the version of the
build-storybook command. If that works, we know we should continue using
build-storybook. If not, we'll use `storybook build`.