Drop support for Storybook v5
Version 6 was released almost four years ago. It's starting to be
hard to test this old version, and to stay leaner when moving forward
I've decided to drop support for Storybook v5 at this point.

We can still maintain a v5 branch if the need for it arises. But
starting with the next release, users can't expect to be able to use
happo-plugin-storybook with Storybook <v6.