Listen to `completed` storyRenderPhaseChanged event by default
And allow people to disable the behavior by calling
`setShouldWaitForCompletedEvent(false)`.

At first I was trying to find a feature-detection method of knowing when
this event was present or not. But it turned out that was quite hard. So
I went the simpler route by defaulting the variable to true and then
allowing people to override it in case they are on Storybook earlier
than v6.4.0.

The issue with the previous feature detection approach was that we
relied on a listener for the event to be set up elsewhere. If no
listener were in place for storyRenderPhaseChanged, we wouldn't use it.
In most cases, the listener is there, but we've just come across
a Storybook where that's not the case.

Fixes https://github.com/happo/happo-plugin-storybook/issues/105