Fall back to using node_modues/.bin script
If `storybook` is defined as a script in package.json, we can't trust
that it can be used to build the storybook package. I'm adding some code
that will fall back to using the binary installed by yarn/npm instead if
we detect a storybook script.

Relying on the .bin/ script isn't 100% safe, so I'm only using this as a
fallback.

Fixes https://github.com/happo/happo-plugin-storybook/issues/108