Add support for `only` init config
This commit is in preparation of a feature coming soon to happo.io.
Right now you can retry a full snap-request, but soon you'll be able to
retry a single snapshot in a full suite. To do that, we need this
library to honor a new `only` configuration option. We use it to single
out the story that we need to render.