Prevent infinite loop when error is thrown during `play` invocation
In case the play function throws an error, we didn't properly
disambiguate it from the internal error thrown by
`forceHappoScreenshot`. This caused the test suite to re-render the last
step over and over (until eventually the browser crashed). To fix this,
I'm adding a `done` boolean to the step created by forceHappoScreenshot.
Each step is only allowed to be processed once.

It's a little annoying that the `errored` event details do not contain
any information about the thrown error. If it did, we wouldn't have to
do this complicated detection to see if the error is ours or someone
elses.