Stop using babel-plugin-transform-runtime
We received a bug report with an error message like this:

```
TypeError: _interopRequireDefault2 is not a function
  at node_modules/happo-plugin-storybook/register.js ...
```

I'm not entirely sure why this is happening, but I believe there might
be some sort of incompatible confluence with the way we are building
this package and the way that @babel/runtime is being installed in the
repo where this error is appearing.

Since babel-plugin-transform-runtime is used primarily to reduce bundle
sizes, and since bundle sizes are not very important in this particular
case, I think we can simplify things and avoid these issues by building
this package without the runtime transform. This also allows us to drop
the dependency on @babel/runtime, which is a nice win.