Update .gitignore and .npmignore
I noticed that there were some large files in the .yarn directory in the
published package that we don't need to distribute in this package.

See here:

https://unpkg.com/browse/happo-plugin-storybook@4.3.0/

I also ended up with some .pnp.* files after getting yarn working in
this project, and I think we want to .gitignore these as well.