Add HAPPO_API_SECRET to storybook6 build
I noticed that the storybook 6 build failed on a default branch build
here:

https://github.com/happo/happo-plugin-storybook/actions/runs/9898839411/job/27346506331#step:11:23

In the logs, I see this:

> No `apiKey` or `apiSecret` found in config. Falling back to
> pull-request authentication.

It looks like the storybook 7 and 8 builds here don't have this issue,
and I also see that those builds have the HAPPO_API_SECRET env var
provided, so I am hoping that this will resolve this problem.

Separately, I think we can improve the error message in this validation
a little bit.