Stabilize "Misc Failing" diffs in Happo
Each time this runs we get new diffs because the port changes, or the
asset hash changes. Example:

https://happo.io/a/77/p/1772/compare/c6918810f0c96cccda1168ee5f1cac9463a275cb/8c6016048861e7a1924277405b8c79a09d79fc45

I think we can stabilize this by using an error boundary and sanitizing
the error stack trace. This makes the story a little less accurate, but
I think that's totally okay in this case.