Reproduce missing interactions bug
I've found an issue when the first story has interactions and you are
using `forceHappoScreenshot`. The check for whether happo is running or
not uses `if (currentIndex)` but the first step is 0. I have a fix in
mind for this but I want to make sure that we have a way to reproduce
first.

A small change to how we order components helped make the Interactive -
Demo [white] example render first.