Add Playwright test for invoking param function
We want to make sure that the fix in f5fd1c8 doesn't regress. The best
idea for a test we could come up with was to use Playwright to load the
storybook and verify that the right things happened on click.

As part of this change I'm switching us back to yarn 1. While being on
newer yarn is generally a good idea, our other repos are all on yarn 1
and it's annoying when switching.