Show help text for function params in addon panel
If core.channelOptions.allowFunctions is false (the default), function
params (like beforeScreenshot, waitFor) will not show up in the Happo
panel. To help users discover this we can show a small help message in
the panel.

I couldn't find an easy way to get the configuration value, so I'm
displaying it all the time.