Check out HEAD of PR instead of merge commit in GitHub actions
We recently made a change to the github action integration so that it
would find the correct merge base. This needs to be able to find
`origin/main`, which doesn't happen when using the checkout action
without the correct ref. This commit updates our checkout in this repo
to match what we recommend in our continuous integration documentation.