Add support for storybook 9, drop support for older versions
In Storybook v9, some packages moved into the main storybook package. We
could probably make this work by doing a big refactor to use dependency
injection via separate entry points, but in the interest of getting this
release out as quickly as possible, we decided to move forward with a
major release that drops support for older storybook versions. We can
backport any critical fixes as needed.

It also looks like Storybook 9 dropped support for the `allowFunction`
option entirely:

https://github.com/storybookjs/storybook/pull/30998

This means we need to remove support for surfacing functions in the
addons panel. From what I can tell, the beforeScreenshot and
afterScreenshot hooks continue to work correctly.