Bring back function params in addons panel
With the Storybook 9 update, the ability to read params with a function
as a value in the addons panel was removed. With that, the
`core.channelOptions.allowFunction: true` was also removed.

In this commit I'm using a workaround to get the params back to the
panel. I'm using the addons channel when mounting a story to communicate
the list of params with functions. I then use that list to populate the
list of params in the addon panel. The same invocation pattern as
before -- sending a message across the addons channel.

To make invoking functions slightly better, I've added some logging of
the return value. If the function returns a promise, we will also await
the result and log them.