Use finished render phase to determine completed story
I've been debugging a test suite with many play functions and
forceHappoScreenshot calls. After they upgraded to Storybook 9 they
startyed seeing cross-contamination, missing snapshots and flakiness. It
turns out that most (all?) flake was coming from the fact that we used
the `completed` and `errored` render phase events to determine that the
story in its current state was done.

Stopping at the error or completed event meant that other events could
end up in our lap later in the cycle. Things as `afterEach` and
`aborted` for instance.

I spent a long time trying to clean up things when the next step in the
play function is supposed to get rendered. In the end I found that
switching to `finished` was much more stable.