Wait for finished even when story is double-rendered
In some cases, a story might re-render itself as part of its rendering
cycle. This happens for instance if you are using the pseudo-states
addon, when it syncs params to globals:

https://github.com/storybookjs/storybook/blob/37c7ef821ba87b739f6dc7ab15cac3f37cea6780/code/addons/pseudo-states/src/preview/withPseudoState.ts#L162

If we don't take this into account, there's a chance that Happo
considers things to be done, but the component continues to change
things in the background. In the pseudo-addons case, we are dealing with
globals which means that the effects a race-condition here might have
are very confusing. And that's exactly what we are seeing in one Happo
user's test suite:
- Theming is flaky, sometimes we see a dark background, sometimes
  there's a light background
- Hover states are applied to the wrong story
- Stories suddenly render nothing

To fix this, I'm keeping count of how many loading phases we have seen
for a story. When we're still waiting for finished phases, we wait.

My manual testing has shown a good improvement as part of making this
change.